Macro hal_x86_64::write_control_reg

macro_rules! write_control_reg {
    ($reg: ident, $value: expr) => { ... };
}

Write to a control register. Calls to this macro will need to be inside an unsafe block, which is intended because writing to control registers can be kinda dangerous. The name of the control register should be passed as any of: CR0, CR1, CR2, CR3, CR4, CR8.