Module hal_x86_64::paging

Structs

Entry

EntryFlags

MapperImpl

PageTable

Table

Enums

Level1

Level2

Level3

Level4

Traits

HierarchicalLevel
Tables of levels that implement HierarchicalLevel are page tables whose entries are other tables, as opposed to actual frames (like in P1s). This makes accessing the next level type-safe, as the next_table methods are only implemented for tables that have child tables.

TableLevel