Struct hal_x86_64::paging::PageTableImpl

pub struct PageTableImpl { /* fields omitted */ }

Methods

impl PageTableImpl

pub fn new(p4_frame: Frame, physical_base: VirtualAddress) -> PageTableImpl

pub unsafe fn from_frame(
p4_frame: Frame,
physical_base: VirtualAddress
) -> PageTableImpl

Create a PageTableImpl from a Frame that already contains a P4. This is very unsafe because it assumes that the frame contains a valid page table, and that no other PageTableImpls currently exist that use this same backing frame (as calling mapper on both could lead to two mutable references aliasing the same data to exist, which is UB).